requirements.txt
numpy
pandas
matplotlib
seaborn
tensorflow-gpu
click
nltk
sacremoses
gensim
lightgbm
sklearn
